Collapse FakeOptionsBase<T> and FakeOptions<T>
